Sunday refactor

+++ client/src/components/Lobby/LobbyWrapper.js
@@ -2,19 +2,30 @@ import React from "react";
 import { useHistory, useParams } from "react-router-dom";
 import { CurrentUserContext } from "../../CurrentUserContext";
 import { LobbyContext } from "../../LobbyContext";
+import { GameRoomContext } from "../../GameRoomContext";
+import Spinner from "../Spinner/Spinner";
 import PlayerHandler from "../SignIn/PlayerHandler";
 import SignIn from "../SignIn/SignIn";
 import HostLobby from "../Lobby/HostLobby";
-import FourOFour from "../FourOFour/FourOFour";
+import ErrorScreen from "../ErrorScreen/ErrorScreen";
+import GameInProgress from "../ErrorScreen/GameInProgress";
 import * as firebase from "firebase";
 
 const LobbyWrapper = () => {
-  const { currentUser, setCurrentUser } = React.useContext(CurrentUserContext);
-  const { setRoomId, roomExists, setRoomExists, location } = React.useContext(
-    LobbyContext
-  );
-
-  console.log("currentUser", currentUser);
+  const {
+    currentUser,
+    setCurrentUser,
+    isHostPresent,
+    setCurrentRoomId,
+  } = React.useContext(CurrentUserContext);
+  const {
+    setRoomId,
+    roomExists,
+    setRoomExists,
+    location,
+    deletePlaylist,
+  } = React.useContext(LobbyContext);
+  const { setGameStarted, setHistoryArray } = React.useContext(GameRoomContext);
 
   const { roomId } = useParams();
   const history = useHistory();
@@ -25,10 +36,15 @@ const LobbyWrapper = () => {
     roomRef.on("value", (snapshot) => {
       setRoomExists(snapshot.exists());
     });
+
+    return () => {
+      roomRef.off();
+    };
   }, [roomId]);
 
   React.useEffect(() => {
     setRoomId(roomId);
+    setCurrentRoomId(roomId);
   }, [roomId]);
 
   React.useEffect(() => {
@@ -36,12 +52,15 @@ const LobbyWrapper = () => {
 
     return () => {
       console.log(history.action);
-      if (history.action === "POP" && location === "lobby")
+      if (history.action === "POP" && location === "lobby" && currentUser)
         if (
           window.confirm("do you really want to leave ?") &&
           currentUser !== null
         ) {
           setCurrentUser(null);
+          setGameStarted(false);
+          deletePlaylist();
+          setHistoryArray([]);
           fetch("/deleteUser", {
             method: "DELETE",
             body: JSON.stringify({ currentUser, roomId }),
@@ -52,6 +71,7 @@ const LobbyWrapper = () => {
           })
             .then((res) => res.json())
             .then((data) => console.log(data));
+          history.push("/");
           return;
         } else if (location === "lobby") {
           console.log("history push from Lobby");
@@ -61,10 +81,35 @@ const LobbyWrapper = () => {
   }, [history, location]);
 
   if (roomExists === false) {
-    return <FourOFour />;
+    return (
+      <ErrorScreen
+        title="404 - Room not found"
+        message="We're sorry, we couldn't find the room you are trying to join"
+      />
+    );
   }
-
-  if (!currentUser && roomExists === true) {
+  if (isHostPresent === false) {
+    return (
+      <ErrorScreen
+        title="The host has left"
+        message="We're sorry, the host has left and the game has ended. Return to the homepage to start a new game."
+      />
+    );
+  }
+  if (
+    !currentUser &&
+    roomExists === true &&
+    isHostPresent === true &&
+    location === "gameRoom"
+  ) {
+    return <GameInProgress />;
+  }
+  if (
+    !currentUser &&
+    roomExists === true &&
+    isHostPresent === true &&
+    location === "lobby"
+  ) {
     return (
       <SignIn
         buttonHandler={PlayerHandler}
@@ -72,7 +117,16 @@ const LobbyWrapper = () => {
       />
     );
   }
-  return <>{roomExists === true && <HostLobby />}</>;
+
+  return (
+    <>
+      {roomExists === true && isHostPresent === true && currentUser ? (
+        <HostLobby />
+      ) : (
+        <Spinner size={50} color={"lightgrey"} />
+      )}
+    </>
+  );
 };
 
 export default LobbyWrapper;
